Adding docs/test_case.md and gofmt
